refactor: move database access function to adhere to sla
