+++ backend/core/db/repositories/cong_repo.go
@@ -1,24 +1,23 @@
-package common
-
-// there's probably a better place for this file but i suck at system design so whatever
-// deal with it :)
-// nah but fr i need help where do i put this??? 😭😭😭😭
+package repositories
 
 import (
 	"backend/core/db/models"
+	"backend/core/services/security"
 
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
 )
 
-func GetCurrentUser(ctx *gin.Context, userId string) *models.User {
+func GetCurrentUser(ctx *gin.Context) *models.User {
+
 	db, _ := ctx.MustGet("db").(*gorm.DB)
 
+	tokenPayload, _ := ctx.MustGet("jwtPayload").(*security.SessionTokenPayload)
 	var foundUser models.User
 	userQueryResult := db.
 		// Preload the congregation to ensure it's included in the user payload
 		Preload("Congregation").
-		First(&foundUser, "id = ?", userId)
+		First(&foundUser, "id = ?", tokenPayload.UserID)
 
 	if userQueryResult.Error != nil {
 		return nil
